refactor: move normalize_column_names_to_schema_casing to schema module
Move the function from create_table.rs to schema/mod.rs since it is a
general-purpose schema utility not specific to CREATE TABLE. ALTER TABLE
and other operations will need the same case normalization logic.
